// generated/api/models/agreementStatus.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * WolfBankGateway
 * OpenAPI spec version: 1.0
 */

export type AgreementStatus = (typeof AgreementStatus)[keyof typeof AgreementStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AgreementStatus = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2
} as const;