// generated/api/models/bankAccountStatus.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * WolfBankGateway
 * OpenAPI spec version: 1.0
 */

export type BankAccountStatus = (typeof BankAccountStatus)[keyof typeof BankAccountStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const BankAccountStatus = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2
} as const;