// generated/api/models/bankAccountType.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * WolfBankGateway
 * OpenAPI spec version: 1.0
 */

export type BankAccountType = (typeof BankAccountType)[keyof typeof BankAccountType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const BankAccountType = {
  NUMBER_0: 0,
  NUMBER_1: 1
} as const;